Conversion fixes
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@106124 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/gd/gd.c
@@ -3988,7 +3988,7 @@ static void php_image_filter_smooth(INTERNAL_FUNCTION_PARAMETERS)
 	if (im_src==NULL) {
 		RETURN_FALSE;
 	}
-	if (gdImageSmooth(im_src, weight)==1) {
+	if (gdImageSmooth(im_src, (float)weight)==1) {
 		RETURN_TRUE;
 	}
 	RETURN_FALSE;
